+++ include/core/rasterizer/PerspectiveCamera.hpp
@@ -8,6 +8,9 @@ public:
     PerspectiveCamera(int screenWidth, int screenHeight, float fov);
 
     IVector2 Project(FVector3 worldPoint) override;
+    IVector2 ProjectTopLeft2D(FVector2 worldPoint, float z = 0.0f) override;
+    float GetProjectedDepth(FVector3 worldPoint) override;
+    Rectangle ProjectRectangleTopLeft(FVector2 position, FVector2 size, float z = 0.0f) override;
     FVector3 WorldToCameraSpace(FVector3 worldPoint) override;
     FVector3 GetPosition() const override;
     void SetPosition(FVector3 position) override;